adicionando interface para services
